finished buy award flow

+++ Code/ServerCode/web_service/ticket/cancel_user_ticket.php
@@ -27,9 +27,45 @@ if($username == ""){
     die();
 }
 
+$query = "select * from customer_tickets where id='".$ticketId."'";
+$query_exec = mysqli_query($localhost, $query);
+$ticket_row = mysqli_fetch_array($query_exec);
+
+if($ticket_row['id'] == ""){
+    echo '{"error":"ticket is not exist"}';
+    die();
+}
+
+$query = "select * from award where id='".$ticket_row['awardID']."'";
+$query_exec = mysqli_query($localhost, $query);
+$award_row = mysqli_fetch_array($query_exec);
+
+if($award_row['id'] == ""){
+    echo '{"error":"award is not exist"}';
+    die();
+}
+
+$query = "select * from customer_shop where username='".$ticket_row['username']."' and shop_id='".$ticket_row['shopID']."'";
+$query_exec = mysqli_query($localhost, $query);
+$customer_shop_row = mysqli_fetch_array($query_exec);
+
+if($customer_shop_row['username'] == ""){
+    echo '{"error":"user not follow this shop"}';
+    die();
+}
+
 $query = "delete from customer_tickets where id='".$ticketId."'";
 $query_exec = mysqli_query($localhost, $query);
 if($query_exec){
+    //update award quantity
+    $new_quantity = $award_row['quantity'] + $ticket_row['quantity'];
+    $query = "update award set quantity = " . $new_quantity . " where id = '".$award_row['id']."'";
+    $query_exec = mysqli_query($localhost, $query);
+    //update point after cancel.
+    $new_point = $customer_shop_row['point'] + $ticket_row['total_point'];
+    $query = "update customer_shop set point = " . $new_point . " where username = '".$ticket_row['username']."' and shop_id = '".$ticket_row['shopID']."'";
+    $query_exec = mysqli_query($localhost, $query);
+
     echo '{"error":""}';
     // Gửi notification cho user
     // Từ $customerName -> regID (bảng customer_registration)
@@ -42,14 +78,14 @@ if($query_exec){
     if($regID != "") {
 
         $regID = array($regID);
-        $message = "trade successfully";
+        $message = "cancel successfully";
         $message = array("message" => $message, "shopID" => $shopId);
 
         $gcm = new GCM();
 
         $result = $gcm->send_notification($regID, $message);
 
-    }
+        }
 }else{
     echo '{"error":"cannot delete this ticket"}';
 }
